BULK-22, #398: Add support for reading from AWS S3 URLs
Motivation:

In order to efficiently upload a large number of records from AWS S3 to
Cassandra, dsbulk needs to be able to read from s3:// URLs.

Modifications:

The main change here is adding the S3URLStreamHandler(Provider) classes.
This makes reading from S3 URLs transparent to the rest of the
application in the same way that reading from stdin/stdout is. A
modifcation to DataStaxBulkLoader was required to be able to pass the
config into these new classes.

Result:

Users may now pass s3:// URLs to load commands and dsbulk will retrieve
the content from AWS S3 and load it into Cassandra.

Future Work:

For this use case, only reading from S3 was required, but adding support
for writing to S3 should be a fairly simple modification to
S3URLStreamHandler.getOutputStream().